Update magic link code
