+++ package.json
@@ -64,7 +64,9 @@
     "react-native-web": "~0.19.13",
     "react-native-webview": "13.12.5",
     "tailwindcss": "^3.4.17",
-    "zod": "^3.24.1"
+    "zod": "^3.24.1",
+    "expo-auth-session": "~6.0.1",
+    "expo-crypto": "~14.0.1"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
